# FileStream: seek only when length exceeds buffer size